+++ sample/src/main/java/com/canhub/cropper/sample/SMainActivity.kt
@@ -1,29 +1,26 @@
 package com.canhub.cropper.sample
 
 import android.os.Bundle
-import android.view.View
 import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
 import com.canhub.cropper.sample.crop_image.app.SCropImageFragment
 import com.canhub.cropper.sample.crop_image_java.app.SCropImageFragmentJava
 import com.canhub.cropper.sample.crop_image_view.app.SCropImageViewFragment
 import com.canhub.cropper.sample.extend_activity.app.SExtendActivity
-import com.example.croppersample.R
 import com.example.croppersample.databinding.ActivityMainBinding
 
 internal class SMainActivity : AppCompatActivity() {
 
+    private lateinit var binding: ActivityMainBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        val binding = ActivityMainBinding.inflate(layoutInflater)
+        binding = ActivityMainBinding.inflate(layoutInflater)
 
         setContentView(binding.root)
 
         binding.sampleCropImageView.setOnClickListener {
-            hideButtons(binding)
-            supportFragmentManager
-                .beginTransaction()
-                .replace(R.id.container, SCropImageViewFragment.newInstance())
-                .commit()
+            SCropImageViewFragment.newInstance().show()
         }
 
         binding.sampleCustomActivity.setOnClickListener {
@@ -31,24 +28,26 @@ internal class SMainActivity : AppCompatActivity() {
         }
 
         binding.sampleCropImage.setOnClickListener {
-            hideButtons(binding)
-            supportFragmentManager
-                .beginTransaction()
-                .replace(R.id.container, SCropImageFragment.newInstance())
-                .commit()
+            SCropImageFragment.newInstance().show()
         }
 
         binding.sampleCropImageJava.setOnClickListener {
-            hideButtons(binding)
-            supportFragmentManager
-                .beginTransaction()
-                .replace(R.id.container, SCropImageFragmentJava.newInstance())
-                .commit()
+            SCropImageFragmentJava.newInstance().show()
         }
     }
 
-    private fun hideButtons(binding: ActivityMainBinding) {
-        binding.image.visibility = View.GONE
-        binding.buttons.visibility = View.GONE
+    private fun Fragment.show() {
+        supportFragmentManager
+            .beginTransaction()
+            .replace(binding.container.id, this)
+            .commit()
+    }
+
+    override fun onBackPressed() {
+        supportFragmentManager.findFragmentById(binding.container.id)?.apply {
+            supportFragmentManager.beginTransaction().remove(this).commit()
+            return
+        }
+        super.onBackPressed()
     }
 }